Plug a state-close-down hole in etorrent_dht.
When we terminate, we should store the state of the DHT table upon normal and shutdowns.
This is achieved by trapping exits and then handling the case explicitly in a nice way in the
code base. I have verified that there are no other linked processes to this process by inspection,
so we don't expect any other kind of 'EXIT' messages to show up.